Add files via upload
This is a modified version of 'read.math' dataset obtained from 'sirt' package. The data contains binary item responses from 664 students to 30 Mathematics items for German fourth graders. Items are classified into three domains: Arithmetic (Items 1-16), Measurement (Items 17-23), and Geometry (Items 24-30). 

Usage
data(data.math)
Format
The dataset is a list. The list element data contains the dataset with the demographic variables student ID (idstud) and a dummy variable for female students (female). The remaining variables (starting with M in the name) are the mathematics items. 
The item metadata are included in the list element item which contains item name (item) and the testlet label (testlet). An item not included in a testlet is indicated by NA. Each item is allocated to one and only competence domain (domain).