Add files via upload
This data contains item responses from 1000 individuals on 8 items measuring science self-efficacy in the Program for International Student Assessment (PISA). The items are measured on a 4-point Likert type rating scale (i.e., polytomously scored). Below are the statements for the eight items.

Recognize the science question that underlies a newspaper report on a health issue.
Explain why earthquakes occur more frequently in some areas than in others.
Describe the role of antibiotics in the treatment of disease.
Identify the science question associated with the disposal of garbage.
Predict how changes to an environment will affect the survival of certain species.
Interpret the scientific information provided on the labeling of food items.
Discuss how new evidence can lead you to change your understanding about th possibility of the life on Mars.
Identify the better of two explanations for the formation of acid rain.

Response scale:
1 = “I couldn’t do this”
2 = “I would struggle to do this on my own”
3 = “I could do this with a bit of effort”
4 = “I could do this easily”