Add files via upload
This is a subset of the data collected through a survey from the Trends in International Mathematics and Science Study (TIMSS) 2015 (for Grade 8). The data contains item responses from 1,000 randomly sampled respondents from the U.S. administration of the 9 items for the Mathematics Self-Efficacy scale. Each item was rated using four response categories: disagree a lot, disagree a little, agree a little, and agree a lot.